Moves Card to clients folder
